+++ src/pages/MyPatches.tsx
@@ -21,7 +21,7 @@ import { useQuery } from "@apollo/react-hooks";
 import { useFilterInputChangeHandler, usePrevious } from "hooks";
 import styled from "@emotion/styled";
 import get from "lodash/get";
-import { PatchCard } from "./my-patches/patch-card-list/PatchCard";
+import { PatchCard } from "./my-patches/PatchCard";
 import { Skeleton, Pagination } from "antd";
 
 export const MyPatches = () => {
