+++ src/pages/my-patches/patch-card/BuildStatusIcon/FailedIcon.tsx
